Added Basic Movement Testing
The most important thing about this was completing the test player's
game object (putting it together at least). Just as well as setting the
center position average and slerping the ship between its point and the
center point. Also, changed the lightning color from blue to pink by
adjusting the tint in the material's parameter thing on the editor.